requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
flask-limiter==3.5.0
pytest==8.3.4
aiohttp==3.9.1
asyncio-throttle==1.0.2
